chore: UnsopportedGateError in error file
